requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
requests>=2.31
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
passlib[bcrypt]>=1.7.4
bcrypt<4.0
PyJWT>=2.8
email-validator>=2.1
